let ID be auto-generated so that the db sequence is at a valid state when test tries to save new entities
out of the box the tests failed for me, DataIntegrityViolationException because of primary key. Looks like the primary key sequence was at 0 when tests started so it tried to give ID=1 to the entity of my test but since V2_populate_sample_data already used this ID, it came to a collision. At least this is my best explanation for why it happened